// apps/enterprise-service-management/src/generated/models/field-definition-model.ts
export const FieldDefinitionFieldTypeKeyToLabel = {
  'Text': 'Text',
  'MultiLineText': 'Multi-line text',
  'Number': 'Number',
  'Date': 'Date',
  'DateAndTime': 'Date and time',
  'YesOrNo': 'Yes or no',
  'SingleChoice': 'Single choice',
  'MultipleChoice': 'Multiple choice',
  'Person': 'Person',
  'Department': 'Department',
  'Site': 'Site',
  'Asset': 'Asset',
  'Attachment': 'Attachment',
  'Email': 'Email',
  'Phone': 'Phone',
  'URL': 'URL',
  'Information': 'Information',
  'Currency': 'Currency',
  'RichText': 'RichText',
  'BusinessUnit': 'BusinessUnit',
  'Lookup': 'Lookup',
  'CascadingChoice': 'CascadingChoice',
  'DisplayOnly': 'DisplayOnly'
} as const;
export type FieldDefinitionFieldTypeKey = keyof typeof FieldDefinitionFieldTypeKeyToLabel;

export const FieldDefinitionWidthKeyToLabel = {
  'Full': 'Full',
  'Half': 'Half',
  'Third': 'Third',
  'TwoThirds': 'TwoThirds'
} as const;
export type FieldDefinitionWidthKey = keyof typeof FieldDefinitionWidthKeyToLabel;

export interface FieldDefinition {
  /**
   * @displayName Field Definition ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  label: string;
  /**
   * @displayName Agent Only
   * @validationRule Required for create/update operations
   */
  agentOnly: boolean;
  /**
   * @displayName Configuration
   */
  configuration?: string;
  /**
   * @displayName Created At
   */
  createdAt?: string;
  /**
   * @displayName Default Value
   */
  defaultValue?: string;
  /**
   * @displayName Field Code
   * @validationRule Required for create/update operations
   */
  fieldCode: string;
  /**
   * @displayName Field Type
   * @validationRule Required for create/update operations
   */
  fieldTypeKey: FieldDefinitionFieldTypeKey;
  /**
   * @displayName Form Definition
   * @validationRule Required for create/update operations
   */
  formDefinition: Pick<FormDefinition, 'id' | 'name1'>;
  /**
   * @displayName Form Section
   * @validationRule Required for create/update operations
   */
  formSection: Pick<FormSection, 'id' | 'title'>;
  /**
   * @displayName Form Version ID
   * @validationRule Required for create/update operations
   */
  formVersionID: Pick<FormVersion, 'id' | 'snapshotLabel'>;
  /**
   * @displayName Help Text
   */
  helpText?: string;
  /**
   * @displayName Is Deleted
   */
  isDeleted?: boolean;
  /**
   * @displayName Placeholder
   */
  placeholder?: string;
  /**
   * @displayName Read Only
   * @validationRule Required for create/update operations
   */
  readOnly1: boolean;
  /**
   * @displayName Reportable
   */
  reportable?: boolean;
  /**
   * @displayName Requester Visible
   * @validationRule Required for create/update operations
   */
  requesterVisible: boolean;
  /**
   * @displayName Required
   * @validationRule Required for create/update operations
   */
  required: boolean;
  /**
   * @displayName Searchable
   */
  searchable?: boolean;
  /**
   * @displayName Sensitive
   * @validationRule Required for create/update operations
   */
  sensitive: boolean;
  /**
   * @displayName Sort Order
   */
  sortOrder?: number;
  /**
   * @displayName Updated At
   */
  updatedAt?: string;
  /**
   * @displayName Visibility Role Key
   */
  visibilityRoleKey?: string;
  /**
   * @displayName Width
   */
  widthKey?: FieldDefinitionWidthKey;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}