// apps/enterprise-service-management/src/generated/models/form-version-model.ts
import type { FormDefinition } from './form-definition-model';
import type { Person } from './person-model';
import type { Workspace } from './workspace-model';

export const FormVersionStatusKeyToLabel = {
  'Draft': 'Draft',
  'Published': 'Published',
  'Retired': 'Retired'
} as const;
export type FormVersionStatusKey = keyof typeof FormVersionStatusKeyToLabel;

export interface FormVersion {
  /**
   * @displayName Form Version ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Snapshot Label
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  snapshotLabel: string;
  /**
   * @displayName Configuration Snapshot
   */
  configurationSnapshot?: string;
  /**
   * @displayName Created At
   */
  createdAt?: string;
  /**
   * @displayName Form Definition
   * @validationRule Required for create/update operations
   */
  formDefinition: Pick<FormDefinition, 'id' | 'name1'>;
  /**
   * @displayName Is Deleted
   */
  isDeleted?: boolean;
  /**
   * @displayName Published At
   */
  publishedAt?: string;
  /**
   * @displayName Published By
   */
  publishedBy?: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: FormVersionStatusKey;
  /**
   * @displayName Updated At
   */
  updatedAt?: string;
  /**
   * @displayName Version Number
   * @validationRule Required for create/update operations
   */
  versionNumber: number;
  /**
   * @displayName Workspace
   * @validationRule Required for create/update operations
   */
  workspace: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _FormVersion = 'FormVersion' as const;