// src/generated/models/request-field-value-model.ts
import type { FormVersion } from './form-version-model';
import type { Request } from './request-model';
import type { Person } from './person-model';
import type { Workspace } from './workspace-model';

export const RequestFieldValueValueTypeKeyToLabel = {
  'Text': 'Text',
  'Number': 'Number',
  'Currency': 'Currency',
  'Date': 'Date',
  'DateTime': 'DateTime',
  'Boolean': 'Boolean',
  'SingleChoice': 'SingleChoice',
  'MultipleChoice': 'MultipleChoice',
  'Person': 'Person',
  'Lookup': 'Lookup',
  'File': 'File',
  'Structured': 'Structured'
} as const;
export type RequestFieldValueValueTypeKey = keyof typeof RequestFieldValueValueTypeKeyToLabel;

export interface RequestFieldValue {
  /**
   * @displayName Request Field Value ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName FieldLabel
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  fieldLabel: string;
  /**
   * @displayName Boolean Value
   */
  booleanValue?: boolean;
  /**
   * @displayName Created At
   */
  createdAt?: string;
  /**
   * @displayName Date Value
   */
  dateValue?: string;
  /**
   * @displayName Display Value
   */
  displayValue?: string;
  /**
   * @displayName FieldDefinitionId
   * @validationRule Required for create/update operations
   */
  fieldDefinitionId: string;
  /**
   * @displayName FormVersion
   * @validationRule Required for create/update operations
   */
  formVersion: Pick<FormVersion, 'id' | 'snapshotLabel'>;
  /**
   * @displayName Is Deleted
   */
  isDeleted?: boolean;
  /**
   * @displayName Number Value
   */
  numberValue?: number;
  /**
   * @displayName RequestId
   * @validationRule Required for create/update operations
   */
  requestId: Pick<Request, 'id' | 'requestNumber'>;
  /**
   * @displayName Stable Field Key
   * @validationRule Required for create/update operations
   */
  stableFieldKey: string;
  /**
   * @displayName Structured Value
   */
  structuredValue?: string;
  /**
   * @displayName Text Value
   */
  textValue?: string;
  /**
   * @displayName UpdatedAt
   * @validationRule Required for create/update operations
   */
  updatedAt: string;
  /**
   * @displayName UpdatedBy
   * @validationRule Required for create/update operations
   */
  updatedBy: Pick<Person, 'id' | 'displayName'>;
  /**
   * @displayName Value
   * @validationRule Required for create/update operations
   */
  value: string;
  /**
   * @displayName Value Type
   */
  valueTypeKey?: RequestFieldValueValueTypeKey;
  /**
   * @displayName WorkspaceId
   * @validationRule Required for create/update operations
   */
  workspaceId: Pick<Workspace, 'id' | 'workspaceName'>;
}

export const _RequestFieldValue = 'RequestFieldValue' as const;